[Pdm2-5] Move Input component inside common directory
