+++ src/app/Auth/components/LoginForm.tsx
@@ -1,6 +1,6 @@
 import { Alert, View } from "react-native";
 import React, { useState } from "react";
-import Input from "../../../common/Input";
+import Input from "../../../common/components/Input";
 import Button from "../../../common/components/Button";
 import { authenticateUser } from "../services/authService";
 import { useRouter } from "expo-router";
